requirements.txt
# AI-Powered Legal Document Intelligence Platform - Dependencies

# Core Framework
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# Document Processing
PyPDF2>=3.0.0
pdfplumber==0.9.0
python-docx>=0.8.11
# textract>=1.6.5  # Commented out due to dependency conflict with pdfplumber
PyMuPDF>=1.23.0

# OCR and Image Processing
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python>=4.8.0

# AI/ML Models
torch>=2.0.0
transformers>=4.30.0
spacy>=3.6.0
sentence-transformers>=2.2.0

# Text Processing
nltk>=3.8.0
regex>=2023.6.3
unicodedata2>=15.0.0

# Visualization
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
requests>=2.31.0
tqdm>=4.65.0
pathlib>=1.0.1
typing-extensions>=4.7.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0

# Optional: For production deployment
gunicorn>=21.2.0
uvicorn>=0.23.0

# System dependencies (install via system package manager)
# tesseract-ocr
# tesseract-ocr-hin
# tesseract-ocr-tam
# tesseract-ocr-tel
# tesseract-ocr-ben
# tesseract-ocr-guj
# tesseract-ocr-kan
# tesseract-ocr-mal
# tesseract-ocr-mar
# tesseract-ocr-ori
# tesseract-ocr-pan
# tesseract-ocr-urd
